[DOM-38006] implement black/flake8 code formatter (#132)
* DOM-38006 - initial formatting setup with tests

* DOM-38006 - implemented flake8 and black to examples

* DOM-38006 - implimented flake8 and black on domino dir

* DOM-38006 - update tests formatter with black

* DOM-38006 - updated contants formats and changelog

* DOM-38006 - updated setup to pull versions correctly

* DOM-38006 - added pr templates

* DOM-38006 - updating with pre-commit changes

* DOM-38006 - remove circleci from pre-commit

* DOM-38006 - updated changelogs, airflow in readme, and tests

* DOM-38006 - updated tests formatting

* DOM-38603 - testing pre-commit

* DOM-38603 - updating linting

* testing pre-commits

* DOM-38603 - updated from feedback

* Updated dominodatalab link in PR form


* Updated  checklist in PR form


* DOM-38006 - updated linting

* DOM-38006 - updated for mypy tests

* DOM-38006 - Updated from feedback

* DOM-38006 - update from comment

+++ tests/scripts/test_spark.py
@@ -3,20 +3,20 @@ Test the creation and utilization of a SparkContext.
 """
 
 import pandas as pd
-from pyspark import SparkConf, SparkContext
-from pyspark.sql import SparkSession, SQLContext
+from pyspark import SparkContext
+from pyspark.sql import SQLContext
 
 if __name__ == "__main__":
- 
+
     sc = SparkContext.getOrCreate()
 
     sqlContext = SQLContext(sc)
     spark = sqlContext.sparkSession
-    
-    df = pd.DataFrame([1, 2, 3, 4, 5], columns=['a'])
-    
-    df['a'] = df['a'].astype(int)
-    
+
+    df = pd.DataFrame([1, 2, 3, 4, 5], columns=["a"])
+
+    df["a"] = df["a"].astype(int)
+
     sdf = spark.createDataFrame(df)
-    
+
     sdf.show()
